Move AnkaiosSDK out of src folder
