+++ setup.cfg
@@ -2,7 +2,7 @@
 testpaths = tests
 
 [coverage:run]
-source = src
+source = .
 omit = 
     */__init__.py
     */_protos/*_pb2.py
