+++ tests/WorkloadState/test_workload_substate_enum.py
@@ -13,8 +13,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import pytest
-from src.AnkaiosSDK import WorkloadStateEnum, WorkloadSubStateEnum
-from src.AnkaiosSDK._protos import _ank_base
+from AnkaiosSDK import WorkloadStateEnum, WorkloadSubStateEnum
+from AnkaiosSDK._protos import _ank_base
 
 
 @pytest.mark.parametrize("state, field, expected", [
